core/src/main/java/at/uni_salzburg/cs/cpcc/core/utils/ByteArrayStreamResponse.java
// This code is part of the CPCC-NG project.
//
// Copyright (c) 2014 Clemens Krainer <clemens.krainer@gmail.com>
//
// This program is free software; you can redistribute it and/or modify
// it under the terms of the GNU General Public License as published by
// the Free Software Foundation; either version 2 of the License, or
// (at your option) any later version.
//
// This program is distributed in the hope that it will be useful,
// but WITHOUT ANY WARRANTY; without even the implied warranty of
// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
// GNU General Public License for more details.
//
// You should have received a copy of the GNU General Public License
// along with this program; if not, write to the Free Software Foundation,
// Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.

package at.uni_salzburg.cs.cpcc.core.utils;

import java.io.ByteArrayInputStream;
import java.io.IOException;
import java.io.InputStream;

import org.apache.tapestry5.StreamResponse;
import org.apache.tapestry5.services.Response;

import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;

/**
 * ByteArrayStreamResponse
 */
public class ByteArrayStreamResponse implements StreamResponse
{
    private byte[] data;
    private String contentType;

    /**
     * @param contentType the content type to be reported to the client.
     * @param data the data to be transmitted to the client.
     */
    @SuppressFBWarnings(value = "EI_EXPOSE_REP2", justification = "Exposed on purpose!")
    public ByteArrayStreamResponse(String contentType, byte[] data)
    {
        this.data = data;
        this.contentType = contentType;
    }

    /**
     * {@inheritDoc}
     */
    @Override
    public String getContentType()
    {
        return contentType;
    }

    /**
     * {@inheritDoc}
     */
    @Override
    public InputStream getStream() throws IOException
    {
        return new ByteArrayInputStream(data);
    }

    /**
     * {@inheritDoc}
     */
    @Override
    public void prepareResponse(Response response)
    {
        // Intentionally empty.
    }

}
